+++ modules/security_global/providers.tf
@@ -1,0 +1,10 @@
+terraform {
+  required_version = ">= 1.9"
+  required_providers {
+    # https://registry.terraform.io/providers/hashicorp/aws/latest
+    aws = {
+      source  = "hashicorp/aws"
+      version = "6.0.0-beta3"
+    }
+  }
+}