+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased v0.9.0
 
 - Edit: Improve module sns
+- Add: Module security_global
 - Dependencies: Upgrade provider aws to 6.0.0-beta3
 
 ## 2025-05-21 v0.8.0